ai-core/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
numpy==1.26.4
faiss-cpu==1.8.0.post1
python-dotenv==1.0.1
FlagEmbedding
